pyproject.toml
[tool.poetry]
name = "idp"
version = "0.1.0"
description = ""
authors = ["MinhCao <minh.cnd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.0"
Pillow = "9.3.0"
pdf2image = "1.16.3"
label-studio = "^1.7.0"
pytesseract = "0.3.10"
datasets = "2.14.5"
transformers = "4.34.0"
torch = "2.1.0"
tensorflow-io-gcs-filesystem = ">=0.23.1 <0.32"
tensorflow = "2.12.0"
seqeval = "1.2.2"
accelerate = ">=0.20.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
jupyter = "^1.0.0"
notebook = "6.5.6"
black = "23.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
